feat: expose openDM
